Код разделен на классы
