+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Алексей Тиньков on 12.03.2023.
+//
+
+import Foundation
+
+// для состояния "Результат квиза"
+struct QuizResultsViewModel {
+  let title: String
+  let text: String
+  let buttonText: String
+}
