+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,16 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Алексей Тиньков on 12.03.2023.
+//
+
+import Foundation
+import UIKit
+
+// для состояния "Вопрос задан"
+struct QuizStepViewModel {
+  let image: UIImage
+  let question: String
+  let questionNumber: String
+}
